Rename module to jaxrs-common.
