+++ jaxrs-commons/src/main/java/org/lfenergy/compas/core/jaxrs/JaxrsConstants.java
@@ -1,0 +1,12 @@
+// SPDX-FileCopyrightText: 2021 Alliander N.V.
+//
+// SPDX-License-Identifier: Apache-2.0
+package org.lfenergy.compas.core.jaxrs;
+
+public class JaxrsConstants {
+    JaxrsConstants() {
+        throw new UnsupportedOperationException("JaxrsConstants class");
+    }
+
+    public static final String COMPAS_COMMONS_V1_NS_URI = "https://www.lfenergy.org/compas/commons/v1";
+}
