+++ jaxrs-commons/src/main/java/org/lfenergy/compas/core/jaxrs/model/package-info.java
@@ -1,0 +1,10 @@
+// SPDX-FileCopyrightText: 2021 Alliander N.V.
+//
+// SPDX-License-Identifier: Apache-2.0
+@XmlSchema(xmlns = {@XmlNs(prefix = "jaxrs-commons", namespaceURI = COMPAS_COMMONS_V1_NS_URI)})
+package org.lfenergy.compas.core.jaxrs.model;
+
+import javax.xml.bind.annotation.XmlNs;
+import javax.xml.bind.annotation.XmlSchema;
+
+import static org.lfenergy.compas.core.jaxrs.JaxrsConstants.COMPAS_COMMONS_V1_NS_URI;
