--- rest-commons/src/main/java/org/lfenergy/compas/core/rest/RestConstants.java
@@ -1,12 +1,0 @@
-// SPDX-FileCopyrightText: 2021 Alliander N.V.
-//
-// SPDX-License-Identifier: Apache-2.0
-package org.lfenergy.compas.core.rest;
-
-public class RestConstants {
-    RestConstants() {
-        throw new UnsupportedOperationException("RestConstants class");
-    }
-
-    public static final String COMPAS_REST_V1_NS_URI = "https://www.lfenergy.org/compas/rest/v1";
-}
