--- rest-commons/src/main/java/org/lfenergy/compas/core/rest/model/package-info.java
@@ -1,10 +1,0 @@
-// SPDX-FileCopyrightText: 2021 Alliander N.V.
-//
-// SPDX-License-Identifier: Apache-2.0
-@XmlSchema(xmlns = {@XmlNs(prefix = "rest-commons", namespaceURI = COMPAS_REST_V1_NS_URI)})
-package org.lfenergy.compas.core.rest.model;
-
-import javax.xml.bind.annotation.XmlNs;
-import javax.xml.bind.annotation.XmlSchema;
-
-import static org.lfenergy.compas.core.rest.RestConstants.COMPAS_REST_V1_NS_URI;
